pyproject.toml
[tool.poetry]
name = "environment"
version = "0.1.0"
description = ""
authors = ["dpenney"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyaml = "^23.9.7"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
